backend/requirements.txt
fastapi
uvicorn[standard]
google-cloud-bigquery
python-dotenv
pydantic-settings
